requirements.txt
streamlit
deepface
opencv-python-headless
tf-keras
pillow
